requirements.txt
prettytable
requests
aiohttp
aiofiles
scipy
pillow
numpy

# pip install -r requirement.txt